refactor(assembly): extract forwardKinematics into pure helper
